+++ NEWS.md
@@ -1,5 +1,7 @@
 # iraceplot 1.01
 
+ * New function `ablation_plot()` to plot the result of `irace::ablation()`.
+ 
 # iraceplot 1.0
 
  * Implement all plots that were available in the `irace` package and a few